Created img2img main package and changed to single package interface
