+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=43.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.pytest.ini_options]
-addopts = "--cov=pix2pix --cov=tunit --cov=unit" #include all packages under src directory
+addopts = "--cov=img2img" #include all packages under src directory
 testpaths = [
     "tests",
 ]
