+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = image2image
+name = img2img
 version = 0.0.1
 description = A package for image to image translation using GANs
 long_description = file: README.md
@@ -40,9 +40,7 @@ dev =
     ; flake8>=3.9
 
 [options.package_data]
-pix2pix = py.typed
-tunit = py.typed
-unit = py.typed
+img2img = py.typed
 
 ; [flake8]
 ; max-line-length = 160
