+++ src/img2img/cfg/enums.py
